% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.ltfr.R
\name{contr.ltfr}
\alias{contr.ltfr}
\title{Create a Full Set of Dummy Variables}
\usage{
contr.ltfr(...)
}
\arguments{
\item{...}{Arguments passed to caret::contr.ltfr}
}
\value{
A matrix produced by caret::contr.ltfr
}
\description{
This function should not be used by the user. This package uses the caret
function predict.dummyVars; for technical reasons, in order for this
to work properly we need to import the function contr.ltfr from caret and
export our own function that calls contr.ltfr from caret.
}
