% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftBart.R
\name{Hypers}
\alias{Hypers}
\title{Create hyperparameter object for SoftBart}
\usage{
Hypers(
  X,
  Y,
  group = NULL,
  alpha = 1,
  beta = 2,
  gamma = 0.95,
  k = 2,
  sigma_hat = NULL,
  shape = 1,
  width = 0.1,
  num_tree = 20,
  tau_rate = 10,
  temperature = 1,
  log_prior = NULL,
  zeta = 0,
  normalize_Y = TRUE,
  weights = NULL
)
}
\arguments{
\item{X}{NxP matrix of training data covariates.}

\item{Y}{Nx1 vector of training data response.}

\item{group}{For each column of X, gives the associated group}

\item{alpha}{Parameter of the Dirichlet parameter of the Gibbs prior}

\item{beta}{Parameter penalizing tree depth in the branching process prior}

\item{gamma}{Parameter penalizing new nodes in the branching process prior}

\item{k}{Related to the signal-to-noise ratio, sigma_mu = 0.5 / (sqrt(num_tree) * k). BART defaults to k = 2.}

\item{sigma_hat}{A prior guess at the conditional variance of Y. If not provided, this is estimated empirically by linear regression.}

\item{shape}{Shape parameter for gating probabilities}

\item{width}{Bandwidth of gating probabilities}

\item{num_tree}{Number of trees in the ensemble}

\item{tau_rate}{Bandwidth prior rate, such that tau ~ Exponential(tau_rate)}

\item{temperature}{Inverse of the temperature for tempered likelihood methods; set = 1 unless you know what you are doing}

\item{log_prior}{Log-prior on the number of allowed predictors for the Gibbs prior}

\item{zeta}{If log_prior is not specified, we set the prior proportional to D^-zeta. Defaults to zeta = 0 for a uniform prior on the model size.}
}
\value{
Returns a list containing the function arguments.
}
\description{
Creates a list which holds all the hyperparameters for use with the softbart
command.
}
