% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdsoftbart.R
\name{pdsoftbart}
\alias{pdsoftbart}
\title{Partial dependence plots for SoftBart}
\usage{
pdsoftbart(
  X,
  Y,
  xind = NULL,
  levs = NULL,
  levquants = c(0.05, (1:9)/10, 0.95),
  pl = FALSE,
  plquants = c(0.05, 0.95),
  ...
)
}
\arguments{
\item{X}{Training data covariates}

\item{Y}{training data response}

\item{xind}{variables to create the partial dependence plots for}

\item{levs}{List of levels of the covariates to evaluate at.}

\item{levquants}{Used if levs is not supplied; takes levs to be quantiles of associated predictors.}

\item{pl}{Create a plot?}

\item{plquants}{Quantiles for the partial dependence plot}

\item{...}{Additional arguments passed to softbart or plot.}
}
\value{
Plot methods do not return anything. pdbart and pd2bart return lists with components given below. 

\itemize{
  \item fd: A matrix (i,j) whose value is the ith draw of f_s(x_s) for the 
            jth value of x_s. 'fd' is for 'function draws'.
            For pdfbart2, fd is a single matrix, where the columns correspond 
            to all possible pairs of values for the pair of variables 
            indicated by xind. That is, all possible (x_i, x_j) where x_i is
            a value in the levs component corresponding to the first x and 
            x_j is a value in the levs component corresponding to the second
            one. The first x changes first.
 \item levs: The list of levels used, each component corresponding to a 
 variable. If argument levs was supplied it is unchanged. Otherwise, the 
 levels in levs are constructed using argument levquants.
}

The remaining componets returned in the list are the same as in the value of bart.
}
\description{
Modified version of the pdbart function from the BayesTree package. 
Run softbart at test observations constructed so that a plot can be created 
displaying the effect of a single variable or pair of variables.
}
