% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.softbart_regression}
\alias{predict.softbart_regression}
\title{Predict for SoftBart Regression}
\usage{
\method{predict}{softbart_regression}(object, newdata, iterations = NULL, ...)
}
\arguments{
\item{object}{A softbart_probit object obtained as output of the softbart_regression function}

\item{newdata}{A dataset to construct predictions on}

\item{iterations}{The iterations we want to get predictions on; includes all of iterations including warmup and thinning iterations. Defaults to the saved iterations, running from (num_burn + num_thin):(num_burn + num_thin * num_save)}

\item{...}{Other arguments passed to predict}
}
\value{
A list containing 
\itemize{
  \item mu: samples of mu for each observation, where pnorm(mu) is the success probability.
  \item mu_mean: posterior mean of mu
}
}
\description{
Computes predictions from the softbart regression model for new data.
}
