% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root mean squared error}
\usage{
rmse(y, yhat)
}
\arguments{
\item{y}{the realized outcomes}

\item{yhat}{the predicted outcomes}
}
\value{
Returns the root mean-squared error.
}
\description{
Computes the root mean-squared error between y and yhat, given by 
sqrt(mean((y - yhat)^2)).
}
\examples{

rmse(c(1,1,1), c(1,0,2))
}
