% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftBart.R
\name{softbart}
\alias{softbart}
\title{Fits the SoftBart model}
\usage{
softbart(X, Y, X_test, hypers = NULL, opts = Opts())
}
\arguments{
\item{X}{NxP matrix of training data covariates}

\item{Y}{Nx1 vector of training data responses}

\item{X_test}{NxP matrix of test data covariates}

\item{hypers}{List of hyperparameter values obtained from Hypers function}

\item{opts}{List of MCMC chain settings obtained from Opts function}
}
\value{
Returns a list with the following components:
\itemize{
  \item y_hat_train: fit to the training data for each iteration of the chain
  \item y_hat_test: fit to the testing data for each iteration of the chain
  \item y_hat_train_mean: fit to the training data, averaged over iterations
  \item y_hat_test_mean: fit to the testing data, averaged over iterations
  \item sigma: posterior samples of the error standard deviations
  \item sigma_mu: posterior samples of sigma_mu, the standard deviation of the leaf node parameters
  \item var_counts: posterior sample of group counts, used for variable importance
  \item beta: posterior samples of beta
  \item gamma: posterior samples of gamma
  \item k: posterior samples of k = 0.5 / (sqrt(num_tree) * sigma_mu)
  \item tau_rate: Posterior samples of tau_rate
  \item loglik_train: Log density evaluated for each observation; can be used with loo package to compute LOO-CV statistics
  \item loglik: Sum of loglik_train for each iteration
}
}
\description{
Runs the Markov chain for the softbart model and collects the output
}
