% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.ltfr.R
\name{contr.ltfr}
\alias{contr.ltfr}
\title{Create a Full Set of Dummy Variables}
\usage{
contr.ltfr(...)
}
\arguments{
\item{...}{A list of arguments.}
}
\value{
A matrix produced containing full sets of dummy variables.
}
\description{
Used with \code{dummyVars} in the \pkg{caret} package to create a full set
 of dummy variables (i.e. less than full rank parameterization).
}
