% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SoftBart.R
\name{Opts}
\alias{Opts}
\title{MCMC options for SoftBart}
\usage{
Opts(num_burn = 2500, num_thin = 1, num_save = 2500, num_print = 100,
  update_sigma_mu = TRUE, update_s = TRUE, update_alpha = TRUE,
  update_beta = FALSE, update_gamma = FALSE, update_tau = TRUE,
  update_tau_mean = TRUE)
}
\arguments{
\item{num_burn}{Number of warmup iterations for the chain.}

\item{num_thin}{Thinning interval for the chain.}

\item{num_save}{The number of samples to collect; in total, num_burn + num_save * num_thin iterations are run}

\item{num_print}{Interval for how often to print the chain's progress}

\item{update_sigma_mu}{If true, sigma_mu/k are updated, with a half-Cauchy prior on sigma_mu centered at the initial guess}

\item{update_s}{If true, s is updated using the Dirichlet prior.}

\item{update_alpha}{If true, alpha is updated using a scaled beta prime prior}

\item{update_beta}{If true, beta is updated using a Normal(0,2^2) prior}

\item{update_gamma}{If true, gamma is updated using a Uniform(0.5, 1) prior}
}
\value{
Returns a list containing the function arguments
}
\description{
Creates a list which provides the parameters for running the Markov chain.
}

