\name{SoftBart-package}
\alias{SoftBart-package}
\alias{SoftBart}
\docType{package}
\title{\packageTitle{SoftBart}}
\description{\packageDescription{SoftBart}}
\details{
  Implements the SoftBART model of Linero and Yang (2018). For routine use, most
  users will interact with the functions softbart, Hypers (to specify 
  hyperparameters), and Opts (to specify options for the chain).
}
\author{
  \packageAuthor{SoftBart}
  
  Maintainer: \packageMaintainer{SoftBart}
}
\references{
  See Linero and Yang (2018) Bayesian Regression Tree Ensembles That Adapt to
  Smoothness and Sparsity, Journal of the Royal Statistical Society.
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}
